Support wall-time and ROS-time timers
Add a new method to Node for creating ROS timers.

Refactor timer implementation to account for the two different types of timers.
Deprecate WallTimer interface and WallTimerImpl and replace with unified Timer interface and TimerImpl.
