+++ rcljava/src/test/java/org/ros2/rcljava/SpinTest.java
@@ -30,7 +30,7 @@ import org.ros2.rcljava.executors.Executor;
 import org.ros2.rcljava.executors.SingleThreadedExecutor;
 import org.ros2.rcljava.node.ComposableNode;
 import org.ros2.rcljava.node.Node;
-import org.ros2.rcljava.timer.WallTimer;
+import org.ros2.rcljava.timer.Timer;
 
 public class SpinTest {
   public static class TimerCallback implements Callback {
@@ -74,7 +74,7 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_once_node");
     TimerCallback timerCallback = new TimerCallback(0);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
 
     ComposableNode composableNode = new ComposableNode() {
@@ -94,7 +94,7 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_once_timeout");
     TimerCallback timerCallback = new TimerCallback(0);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
 
     ComposableNode composableNode = new ComposableNode() {
@@ -114,7 +114,7 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_once_timeout_too_short_node");
     TimerCallback timerCallback = new TimerCallback(0);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
 
     ComposableNode composableNode = new ComposableNode() {
@@ -134,7 +134,7 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_some_node");
     TimerCallback timerCallback = new TimerCallback(0);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
 
     ComposableNode composableNode = new ComposableNode() {
@@ -162,7 +162,7 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_some_max_duration_node");
     TimerCallback timerCallback = new TimerCallback(0);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
 
     ComposableNode composableNode = new ComposableNode() {
@@ -190,8 +190,8 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_some_max_duration_multiple_timers_node");
     TimerCallback timerCallback = new TimerCallback(200);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
-    WallTimer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
     assertNotEquals(0, timer2.getHandle());
 
@@ -220,8 +220,8 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_some_max_duration_too_short_node");
     TimerCallback timerCallback = new TimerCallback(200);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
-    WallTimer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
     assertNotEquals(0, timer2.getHandle());
 
@@ -250,7 +250,7 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_all_node");
     TimerCallback timerCallback = new TimerCallback(0);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
 
     ComposableNode composableNode = new ComposableNode() {
@@ -279,8 +279,8 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_all_max_duration_multiple_timers_node");
     TimerCallback timerCallback = new TimerCallback(200);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
-    WallTimer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
     assertNotEquals(0, timer2.getHandle());
 
@@ -309,8 +309,8 @@ public class SpinTest {
     Executor executor = new SingleThreadedExecutor();
     final Node node = RCLJava.createNode("spin_all_max_duration_too_short_node");
     TimerCallback timerCallback = new TimerCallback(200);
-    WallTimer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
-    WallTimer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
+    Timer timer2 = node.createWallTimer(100, TimeUnit.MILLISECONDS, timerCallback);
     assertNotEquals(0, timer.getHandle());
     assertNotEquals(0, timer2.getHandle());
